+++ CHANGELOG.md
@@ -54,6 +54,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Qt 6.11.0 support.
 
+- Full CMake configuration and build system for Visual Studio  (x64, Win32, ARM64).
+
 ### Changed
 
 - Reduced interpreter overhead in tight loops.